+++ NEWS.md
@@ -1,5 +1,7 @@
 # forcats (development version)
 
+* `fct_inorder()` now works when not all levels appear in the data (#262).
+
 * `fct_unique()` now captures implicit missing values if present (#293).
 
 * `fct_other()` no longer generates a warning if no levels are replaced with